add a check to know if the binding is about something inside the current node or something inside one of its child nodes
Indeed, the traces contains everything about reconfiguration of the child nodes.
Maybe, in term of performance, it will be better to fix the traces instead of adding this check but traces are related to kmf and I am not enough confident to look at it to find a better solution.